+++ cli/src/library.rs
@@ -51,7 +51,7 @@ impl<T: ReplLiteral> StdLibrary<T> {
         let constants = self
             .constants
             .iter()
-            .map(|(name, constant)| (*name, Value::Number(constant.to_owned())));
+            .map(|(name, constant)| (*name, Value::Prim(constant.to_owned())));
         let unary_fns = self
             .unary
             .iter()
@@ -160,7 +160,7 @@ where
 
 pub fn create_modular_env(modulus: u64) -> (Environment<'static, u64>, TypeEnvironment) {
     let mut env: Environment<'_, u64> = Prelude.iter().chain(Assertions.iter()).collect();
-    env.insert("MAX_VALUE", Value::Number(modulus - 1));
+    env.insert("MAX_VALUE", Value::Prim(modulus - 1));
 
     let type_env = defs::Prelude::iter()
         .chain(defs::Assertions::iter())